Removed backup/restore stuff
