+++ client_test.go
@@ -15,7 +15,6 @@
 package pilosa_test
 
 import (
-	"bytes"
 	"context"
 	"fmt"
 	"net/http"
@@ -367,133 +366,6 @@ func TestClient_ImportValue(t *testing.T) {
 	}
 	if max != 40 || cnt != 1 {
 		t.Fatalf("unexpected values: got max=%v, count=%v; expected max=40, cnt=1", max, cnt)
-	}
-}
-
-// Ensure client backup and restore a frame.
-func TestClient_BackupRestore(t *testing.T) {
-	hldr := test.MustOpenHolder()
-	defer hldr.Close()
-
-	hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 0).MustSetBits(100, 1, 2, 3, SliceWidth-1)
-	hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 1).MustSetBits(100, SliceWidth, SliceWidth+2)
-	hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 5).MustSetBits(100, (5*SliceWidth)+1)
-	hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 0).MustSetBits(200, 20000)
-
-	s := test.NewServer()
-	defer s.Close()
-	s.Handler.API.Cluster = test.NewCluster(1)
-	s.Handler.API.Cluster.Nodes[0].URI = s.HostURI()
-	s.Handler.API.Holder = hldr.Holder
-
-	c := test.MustNewClient(s.Host(), defaultClient)
-
-	// Backup from frame.
-	var buf bytes.Buffer
-	if err := c.BackupTo(context.Background(), &buf, "i", "f", pilosa.ViewStandard); err != nil {
-		t.Fatal(err)
-	}
-
-	// Restore to a different frame.
-	if _, err := hldr.MustCreateIndexIfNotExists("x", pilosa.IndexOptions{}).CreateFrameIfNotExists("y", pilosa.FrameOptions{}); err != nil {
-		t.Fatal(err)
-	}
-	if err := c.RestoreFrom(context.Background(), &buf, "x", "y", pilosa.ViewStandard); err != nil {
-		t.Fatal(err)
-	}
-
-	// Verify data.
-	if a := hldr.Fragment("x", "y", pilosa.ViewStandard, 0).Row(100).Bits(); !reflect.DeepEqual(a, []uint64{1, 2, 3, SliceWidth - 1}) {
-		t.Fatalf("unexpected bits(0): %+v", a)
-	}
-	if a := hldr.Fragment("x", "y", pilosa.ViewStandard, 1).Row(100).Bits(); !reflect.DeepEqual(a, []uint64{SliceWidth, SliceWidth + 2}) {
-		t.Fatalf("unexpected bits(0): %+v", a)
-	}
-	if a := hldr.Fragment("x", "y", pilosa.ViewStandard, 5).Row(100).Bits(); !reflect.DeepEqual(a, []uint64{(5 * SliceWidth) + 1}) {
-		t.Fatalf("unexpected bits(0): %+v", a)
-	}
-	if a := hldr.Fragment("x", "y", pilosa.ViewStandard, 0).Row(200).Bits(); !reflect.DeepEqual(a, []uint64{20000}) {
-		t.Fatalf("unexpected bits: %+v", a)
-	}
-}
-
-// Ensure client backup and restore a frame with inverse view.
-func TestClient_BackupInverseView(t *testing.T) {
-	hldr := test.MustOpenHolder()
-	defer hldr.Close()
-
-	idx := hldr.MustCreateIndexIfNotExists("i", pilosa.IndexOptions{})
-	frameOpts := pilosa.FrameOptions{
-		InverseEnabled: true,
-	}
-	frame, err := idx.CreateFrameIfNotExists("f", frameOpts)
-	if err != nil {
-		panic(err)
-	}
-	v, err := frame.CreateViewIfNotExists(pilosa.ViewInverse)
-	if err != nil {
-		panic(err)
-	}
-	f, err := v.CreateFragmentIfNotExists(0)
-	if err != nil {
-		panic(err)
-	}
-
-	f.SetBit(100, 1)
-	f.SetBit(100, 2)
-	f.SetBit(100, 3)
-	f.SetBit(100, SliceWidth-1)
-
-	s := test.NewServer()
-	defer s.Close()
-
-	s.Handler.API.Cluster = test.NewCluster(1)
-	s.Handler.API.Cluster.Nodes[0].URI = s.HostURI()
-	s.Handler.API.Holder = hldr.Holder
-
-	c := test.MustNewClient(s.Host(), defaultClient)
-
-	// Backup from frame.
-	var buf bytes.Buffer
-	if err := c.BackupTo(context.Background(), &buf, "i", "f", pilosa.ViewInverse); err != nil {
-		t.Fatal(err)
-	}
-
-	// Restore to a different frame.
-	if _, err := hldr.MustCreateIndexIfNotExists("x", pilosa.IndexOptions{}).CreateFrameIfNotExists("y", pilosa.FrameOptions{InverseEnabled: true}); err != nil {
-		t.Fatal(err)
-	}
-	if err := c.RestoreFrom(context.Background(), &buf, "x", "y", pilosa.ViewInverse); err != nil {
-		t.Fatal(err)
-	}
-
-	// Verify data.
-	if a := hldr.Fragment("x", "y", pilosa.ViewInverse, 0).Row(100).Bits(); !reflect.DeepEqual(a, []uint64{1, 2, 3, SliceWidth - 1}) {
-		t.Fatalf("unexpected bits(0): %+v", a)
-	}
-
-}
-
-// backup returns error with invalid view
-func TestClient_BackupInvalidView(t *testing.T) {
-	hldr := test.MustOpenHolder()
-	defer hldr.Close()
-
-	hldr.MustCreateFragmentIfNotExists("i", "f", pilosa.ViewStandard, 0).MustSetBits(100, 1, 2, 3, SliceWidth-1)
-
-	s := test.NewServer()
-	defer s.Close()
-	s.Handler.API.Cluster = test.NewCluster(1)
-	s.Handler.API.Cluster.Nodes[0].URI = s.HostURI()
-	s.Handler.API.Holder = hldr.Holder
-
-	c := test.MustNewClient(s.Host(), defaultClient)
-
-	// Backup from frame.
-	var buf bytes.Buffer
-	err := c.BackupTo(context.Background(), &buf, "i", "f", "invalid_view")
-	if err != pilosa.ErrInvalidView {
-		t.Fatal(err)
 	}
 }
