+++ ctl/bench_test.go
@@ -15,6 +15,7 @@
 package ctl
 
 import (
+	"bufio"
 	"bytes"
 	"context"
 	"fmt"
@@ -85,4 +86,13 @@ func TestBenchCommand_Run(t *testing.T) {
 	if err != nil {
 		fmt.Println(buf.String())
 	}
+}
+
+// declare stdin, stdout, stderr
+func GetIO(buf bytes.Buffer) (io.Reader, io.Writer, io.Writer) {
+	rder := []byte{}
+	stdin := bytes.NewReader(rder)
+	stdout := bufio.NewWriter(&buf)
+	stderr := bufio.NewWriter(&buf)
+	return stdin, stdout, stderr
 }
