+++ server/server_test.go
@@ -279,80 +279,6 @@ func TestMain_InverseSlices(t *testing.T) {
 	}
 }
 
-// Ensure program can set bits on one cluster and then restore to a second cluster.
-func TestMain_FrameRestore(t *testing.T) {
-	mains1 := test.MustRunMainWithCluster(t, 2)
-	m10 := mains1[0]
-	m11 := mains1[1]
-
-	// Create frames.
-	client := m10.Client()
-	if err := client.CreateIndex(context.Background(), "i", pilosa.IndexOptions{}); err != nil && err != pilosa.ErrIndexExists {
-		t.Fatal("create index:", err)
-	}
-	if err := client.CreateFrame(context.Background(), "i", "f", pilosa.FrameOptions{}); err != nil {
-		t.Fatal("create frame:", err)
-	}
-
-	// Write data on first cluster.
-	if _, err := m10.Query("i", "", `
-		SetBit(row=1, frame="f", col=100)
-		SetBit(row=1, frame="f", col=1000)
-		SetBit(row=1, frame="f", col=100000)
-		SetBit(row=1, frame="f", col=200000)
-		SetBit(row=1, frame="f", col=400000)
-		SetBit(row=1, frame="f", col=600000)
-		SetBit(row=1, frame="f", col=800000)
-	`); err != nil {
-		t.Fatal("setting bits:", err)
-	}
-
-	// Query row on first cluster.
-	if res, err := m10.Query("i", "", `Bitmap(row=1, frame="f")`); err != nil {
-		t.Fatal("bitmap query:", err)
-	} else if res != `{"results":[{"attrs":{},"bits":[100,1000,100000,200000,400000,600000,800000]}]}`+"\n" {
-		t.Fatalf("unexpected result: %s", res)
-	}
-
-	// Start second cluster.
-	mains2 := test.MustRunMainWithCluster(t, 2)
-	m20 := mains2[0]
-	defer m20.Close()
-	m21 := mains2[1]
-	defer m21.Close()
-
-	// Import from first cluster.
-	client20, err := pilosa.NewInternalHTTPClient(m20.Server.URI.HostPort(), server.GetHTTPClient(nil))
-	if err != nil {
-		t.Fatal("new client:", err)
-	}
-	client21, err := pilosa.NewInternalHTTPClient(m21.Server.URI.HostPort(), server.GetHTTPClient(nil))
-	if err != nil {
-		t.Fatal("new client:", err)
-	}
-
-	if err := m20.Client().CreateIndex(context.Background(), "i", pilosa.IndexOptions{}); err != nil && err != pilosa.ErrIndexExists {
-		t.Fatal("create new index:", err)
-	}
-	if err := m20.Client().CreateFrame(context.Background(), "i", "f", pilosa.FrameOptions{}); err != nil {
-		t.Fatal("create new frame:", err)
-	}
-
-	if err := client20.RestoreFrame(context.Background(), m10.Server.URI.HostPort(), "i", "f"); err != nil {
-		t.Fatal("restore frame:", err)
-	}
-	if err := client21.RestoreFrame(context.Background(), m11.Server.URI.HostPort(), "i", "f"); err != nil {
-		t.Fatal("restore frame:", err)
-	}
-
-	// Query row on second cluster.
-	if res, err := m20.Query("i", "", `Bitmap(row=1, frame="f")`); err != nil {
-		t.Fatal("another bitmap query:", err)
-	} else if res != `{"results":[{"attrs":{},"bits":[100,1000,100000,200000,400000,600000,800000]}]}`+"\n" {
-		t.Fatalf("2unexpected result: %s", res)
-	}
-}
-
 // Ensure the host can be parsed.
 func TestConfig_Parse_Host(t *testing.T) {
 	if c, err := ParseConfig(`bind = "local"`); err != nil {
